Allow the user to specify the scrape interval
A default for all jobs can be specified through a cli argument or a config in the am.toml. It is also possible for every endpoint to be overridden using the am.toml.

This also introduces a shorter default scrape interval of 5s if nothing was specified.

Group some cli arguments into groups